auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
python-dotenv
cryptography
bcrypt==3.2.2
